Refactor file upload functions #189
