+++ tests/products.test.js
@@ -1,5 +1,3 @@
-import fetch from 'isomorphic-unfetch';
-import FormData from 'form-data';
 import {
   setupDatabase,
   createLoggedInGraphqlFetch,
@@ -16,8 +14,6 @@ let connection;
 let graphqlFetchAsAdmin;
 let graphqlFetchAsNormalUser;
 let graphqlFetchAsAnonymousUser;
-const fs = require('fs');
-const path = require('path');
 
 describe('Products', () => {
   beforeAll(async () => {
@@ -404,194 +400,6 @@ describe('Products', () => {
       });
 
       expect(errors[0]?.extensions?.code).toEqual('InvalidIdError');
-    });
-  });
-
-  describe('Mutation.addProductMedia for admin user should', () => {
-    it('upload product media correctly', async () => {
-      const body = new FormData();
-      body.append(
-        'operations',
-        JSON.stringify({
-          query: `
-          mutation addProductMedia($productId: ID!, $media: Upload!){
-            addProductMedia(productId: $productId, media: $media){
-              _id
-              tags
-              sortKey
-              file {
-                _id
-                name
-                type
-                url
-              }
-            }
-          }
-        `,
-          variables: {
-            productId: SimpleProduct._id,
-            media: null,
-          },
-        }),
-      );
-
-      body.append('map', JSON.stringify({ 1: ['variables.media'] }));
-      body.append(
-        '1',
-        fs.createReadStream(path.resolve(__dirname, `./assets/image.jpg`)),
-      );
-      const {
-        data: { addProductMedia },
-      } = await fetch('http://localhost:3000/graphql', {
-        method: 'POST',
-        headers: {
-          authorization: ADMIN_TOKEN,
-        },
-        body,
-      }).then((response) => response.json());
-
-      expect(addProductMedia?.file.name).toEqual('image.jpg');
-    });
-
-    it('return ProductNotFoundError when passed non existing product ID', async () => {
-      const body = new FormData();
-      body.append(
-        'operations',
-        JSON.stringify({
-          query: `
-          mutation addProductMedia($productId: ID!, $media: Upload!){
-            addProductMedia(productId: $productId, media: $media){
-              _id
-            }
-          }
-        `,
-          variables: {
-            productId: 'non-existing-id',
-            media: null,
-          },
-        }),
-      );
-
-      body.append('map', JSON.stringify({ 1: ['variables.media'] }));
-      body.append(
-        '1',
-        fs.createReadStream(path.resolve(__dirname, `./assets/image.jpg`)),
-      );
-      const { errors } = await fetch('http://localhost:3000/graphql', {
-        method: 'POST',
-        headers: {
-          authorization: ADMIN_TOKEN,
-        },
-        body,
-      }).then((response) => response.json());
-
-      expect(errors[0]?.extensions?.code).toEqual('ProductNotFoundError');
-    });
-
-    it('return InvalidIdError when passed Invalid product ID', async () => {
-      const body = new FormData();
-      body.append(
-        'operations',
-        JSON.stringify({
-          query: `
-          mutation addProductMedia($productId: ID!, $media: Upload!){
-            addProductMedia(productId: $productId, media: $media){
-              _id
-            }
-          }
-        `,
-          variables: {
-            productId: '',
-            media: null,
-          },
-        }),
-      );
-
-      body.append('map', JSON.stringify({ 1: ['variables.media'] }));
-      body.append(
-        '1',
-        fs.createReadStream(path.resolve(__dirname, `./assets/image.jpg`)),
-      );
-      const { errors } = await fetch('http://localhost:3000/graphql', {
-        method: 'POST',
-        headers: {
-          authorization: ADMIN_TOKEN,
-        },
-        body,
-      }).then((response) => response.json());
-
-      expect(errors[0]?.extensions?.code).toEqual('InvalidIdError');
-    });
-  });
-
-  describe('Mutation.addProductMedia for normal user should', () => {
-    it('return NoPermissionError', async () => {
-      const body = new FormData();
-      body.append(
-        'operations',
-        JSON.stringify({
-          query: `
-          mutation addProductMedia($productId: ID!, $media: Upload!){
-            addProductMedia(productId: $productId, media: $media){
-              _id
-            }
-          }
-        `,
-          variables: {
-            productId: SimpleProduct._id,
-            media: null,
-          },
-        }),
-      );
-
-      body.append('map', JSON.stringify({ 1: ['variables.media'] }));
-      body.append(
-        '1',
-        fs.createReadStream(path.resolve(__dirname, `./assets/image.jpg`)),
-      );
-      const { errors } = await fetch('http://localhost:3000/graphql', {
-        method: 'POST',
-        headers: {
-          authorization: USER_TOKEN,
-        },
-        body,
-      }).then((response) => response.json());
-
-      expect(errors[0]?.extensions?.code).toEqual('NoPermissionError');
-    });
-  });
-
-  describe('Mutation.addProductMedia for anonymous user should', () => {
-    it('return NoPermissionError', async () => {
-      const body = new FormData();
-      body.append(
-        'operations',
-        JSON.stringify({
-          query: `
-          mutation addProductMedia($productId: ID!, $media: Upload!){
-            addProductMedia(productId: $productId, media: $media){
-              _id
-            }
-          }
-        `,
-          variables: {
-            productId: SimpleProduct._id,
-            media: null,
-          },
-        }),
-      );
-
-      body.append('map', JSON.stringify({ 1: ['variables.media'] }));
-      body.append(
-        '1',
-        fs.createReadStream(path.resolve(__dirname, `./assets/image.jpg`)),
-      );
-      const { errors } = await fetch('http://localhost:3000/graphql', {
-        method: 'POST',
-        body,
-      }).then((response) => response.json());
-
-      expect(errors[0]?.extensions?.code).toEqual('NoPermissionError');
     });
   });
